scanner/requirements.txt
# Data & API
polygon-api-client>=1.12.0
requests>=2.31.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# HTML Generation
jinja2>=3.1.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
